# internal/repo/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package repo

import (
	"time"
)

type Application struct {
	ID           int64      `json:"id"`
	JobID        int64      `json:"job_id"`
	UserID       int64      `json:"user_id"`
	Status       string     `json:"status"`
	Notes        *string    `json:"notes"`
	NextActionAt *time.Time `json:"next_action_at"`
	CreatedAt    time.Time  `json:"created_at"`
	UpdatedAt    time.Time  `json:"updated_at"`
}

type Event struct {
	ID            int64     `json:"id"`
	UserID        *int64    `json:"user_id"`
	ApplicationID *int64    `json:"application_id"`
	Type          string    `json:"type"`
	PayloadJson   []byte    `json:"payload_json"`
	CreatedAt     time.Time `json:"created_at"`
}

type Job struct {
	ID        int64     `json:"id"`
	OrgID     *int64    `json:"org_id"`
	Title     string    `json:"title"`
	Company   string    `json:"company"`
	Url       *string   `json:"url"`
	Location  *string   `json:"location"`
	Tags      []string  `json:"tags"`
	CreatedAt time.Time `json:"created_at"`
}

type RefreshToken struct {
	ID        int64      `json:"id"`
	UserID    int64      `json:"user_id"`
	TokenHash string     `json:"token_hash"`
	ExpiresAt time.Time  `json:"expires_at"`
	RevokedAt *time.Time `json:"revoked_at"`
	CreatedAt time.Time  `json:"created_at"`
}

type User struct {
	ID           int64     `json:"id"`
	Email        string    `json:"email"`
	PasswordHash string    `json:"password_hash"`
	CreatedAt    time.Time `json:"created_at"`
}